+++ R/tm_g_scatterplot.R
@@ -236,11 +236,31 @@ ui_g_scatterplot <- function(id, ...) {
           data_extract_spec = args$x,
           is_single_dataset = is_single_dataset_value
         ),
+        checkboxInput(ns("log_x"), "Use log transformation", value = FALSE),
+        conditionalPanel(
+          condition = paste0("input['", ns("log_x"), "'] == true"),
+          radioButtons(
+            ns("log_x_base"),
+            label = NULL,
+            inline = TRUE,
+            choices = c("Natural" = "log", "Base 10" = "log10", "Base 2" = "log2")
+          )
+        ),
         teal.transform::data_extract_ui(
           id = ns("y"),
           label = "Y variable",
           data_extract_spec = args$y,
           is_single_dataset = is_single_dataset_value
+        ),
+        checkboxInput(ns("log_y"), "Use log transformation", value = FALSE),
+        conditionalPanel(
+          condition = paste0("input['", ns("log_y"), "'] == true"),
+          radioButtons(
+            ns("log_y_base"),
+            label = NULL,
+            inline = TRUE,
+            choices = c("Natural" = "log", "Base 10" = "log10", "Base 2" = "log2")
+          )
         ),
         if (!is.null(args$color_by)) {
           teal.transform::data_extract_ui(
@@ -468,6 +488,9 @@ srv_g_scatterplot <- function(id,
       smoothing_degree <- as.integer(input$smoothing_degree)
       ci <- input$ci # nolint
 
+      log_x <- input$log_x
+      log_y <- input$log_y
+
       validate(need(!is.null(ggtheme), "Please select a theme."))
       validate(need(length(x_var) == 1, "There must be exactly one x var."))
       validate(need(length(y_var) == 1, "There must be exactly one y var."))
@@ -496,6 +519,27 @@ srv_g_scatterplot <- function(id,
           "Marginal plots cannot be produced when the points are colored by Date or POSIX variables.
         \n Uncheck the 'Add marginal density' checkbox to display the plot."
         ))
+      }
+
+      if (log_x) {
+        validate(
+          need(
+            is.numeric(ANL[[x_var]]) && all(
+              ANL[[x_var]] > 0 | is.na(ANL[[x_var]])
+            ),
+            "X variable can only be log transformed if variable is numeric and all values are positive."
+          )
+        )
+      }
+      if (log_y) {
+        validate(
+          need(
+            is.numeric(ANL[[y_var]]) && all(
+              ANL[[y_var]] > 0 | is.na(ANL[[y_var]])
+            ),
+            "Y variable can only be log transformed if variable is numeric and all values are positive."
+          )
+        )
       }
 
       teal::validate_has_data(ANL[, c(x_var, y_var)], 10, complete = TRUE, allow_inf = FALSE)
@@ -543,6 +587,36 @@ srv_g_scatterplot <- function(id,
 
       plot_call <- substitute(expr = pre_pro_anl %>% ggplot(), env = list(pre_pro_anl = str2lang(pre_pro_anl)))
 
+      if (log_x) {
+        log_x_fn <- input$log_x_base
+        teal.code::chunks_push(
+          id = "log_x_transformation",
+          expression = substitute(
+            expr = ANL[, log_x_var] <- log_x_fn(ANL[, x_var]), # nolint
+            env = list(
+              x_var = x_var,
+              log_x_fn = as.name(log_x_fn),
+              log_x_var = paste0(log_x_fn, "_", x_var)
+            )
+          )
+        )
+      }
+
+      if (log_y) {
+        log_y_fn <- input$log_y_base
+        teal.code::chunks_push(
+          id = "log_y_transformation",
+          expression = substitute(
+            expr = ANL[, log_y_var] <- log_y_fn(ANL[, y_var]), # nolint
+            env = list(
+              y_var = y_var,
+              log_y_fn = as.name(log_y_fn),
+              log_y_var = paste0(log_y_fn, "_", y_var)
+            )
+          )
+        )
+      }
+
       plot_call <- if (length(color_by_var) == 0) {
         substitute(
           expr = plot_call +
@@ -550,8 +624,8 @@ srv_g_scatterplot <- function(id,
             geom_point(alpha = alpha_value, size = point_sizes, shape = shape_value, color = color_value),
           env = list(
             plot_call = plot_call,
-            x_name = as.name(x_var),
-            y_name = as.name(y_var),
+            x_name = if (log_x) as.name(paste0(log_x_fn, "_", x_var)) else as.name(x_var),
+            y_name = if (log_y) as.name(paste0(log_y_fn, "_", y_var)) else as.name(y_var),
             alpha_value = alpha,
             point_sizes = point_sizes,
             shape_value = shape,
@@ -565,8 +639,8 @@ srv_g_scatterplot <- function(id,
             geom_point(alpha = alpha_value, size = point_sizes, shape = shape_value),
           env = list(
             plot_call = plot_call,
-            x_name = as.name(x_var),
-            y_name = as.name(y_var),
+            x_name = if (log_x) as.name(paste0(log_x_fn, "_", x_var)) else as.name(x_var),
+            y_name = if (log_y) as.name(paste0(log_y_fn, "_", y_var)) else as.name(y_var),
             color_by_var_name = as.name(color_by_var),
             alpha_value = alpha,
             point_sizes = point_sizes,
@@ -690,8 +764,18 @@ srv_g_scatterplot <- function(id,
         plot_call <- substitute(expr = plot_call + facet_cl, env = list(plot_call = plot_call, facet_cl = facet_cl))
       }
 
-      y_label <- varname_w_label(y_var, ANL)
-      x_label <- varname_w_label(x_var, ANL)
+      y_label <- varname_w_label(
+        y_var,
+        ANL,
+        prefix = if (log_y) paste(log_y_fn, "(") else NULL,
+        suffix = if (log_y) ")" else NULL
+      )
+      x_label <- varname_w_label(
+        x_var,
+        ANL,
+        prefix = if (log_x) paste(log_x_fn, "(") else NULL,
+        suffix = if (log_x) ")" else NULL
+      )
 
       dev_ggplot2_args <- teal.widgets::ggplot2_args(
         labs = list(y = y_label, x = x_label),
@@ -761,9 +845,10 @@ srv_g_scatterplot <- function(id,
     )
 
     output$data_table <- DT::renderDataTable({
-      # if not dependent on plot_r() it tries to print a table before chunks are populated with the correct ANL
+      # if not dependent on plot_r() initially it tries to print a table
+      # before chunks are populated with the correct ANL
       # in plot_r which ends up in a bunch of warnings that ANL is not in chunks
-      plot_r()
+      isolate(plot_r())
       plot_brush <- brush$brush()
 
       if (!is.null(plot_brush)) {